requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
mysql==0.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.4
numpy==1.17.4
pandas==0.25.3
protobuf==3.10.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.7
uuid==1.30
